Removed DetectorViewController
